microservices/shared/requirements.txt
appwrite==6.0.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
PyJWT==2.9.0
bcrypt==4.2.0
httpx==0.27.2
redis[asyncio]==5.1.0
groq==0.11.0
boto3==1.35.0
